Move AsyncShutdownable out of internal package
